refactor(SwipeCellDelegate): Created superclass for SwipeTableViewController to inherit all function
