+++ Todoey/Controllers/TodoListViewController.swift
@@ -41,7 +41,7 @@ class TodoListViewController: UITableViewController {
     }
     
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "ToDoItemCell", for: indexPath)
+        let cell = tableView.dequeueReusableCell(withIdentifier: "Cell", for: indexPath)
        
         if let item = todoItems?[indexPath.row] {
             cell.textLabel?.text = item.title
